+++ tests/dataframe/test_aggregations.py
@@ -1,16 +1,11 @@
 from __future__ import annotations
 
-from typing import Any
-
+import pyarrow as pa
 import pytest
 
 from daft import DataFrame, col
+from daft.datatype import DataType
 from daft.errors import ExpressionTypeError
-from tests.conftest import assert_pydict_equals
-
-
-def _sort_list_column_data(list_column_data: list[list[Any]]):
-    return [sorted(l, key=lambda x: (x is None, x)) for l in list_column_data]
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 4])
@@ -28,7 +23,6 @@ def test_agg_global(repartition_nparts):
             (col("values").alias("min"), "min"),
             (col("values").alias("max"), "max"),
             (col("values").alias("count"), "count"),
-            (col("values").alias("list"), "list"),
         ]
     )
     expected = {
@@ -37,18 +31,12 @@ def test_agg_global(repartition_nparts):
         "min": [1],
         "max": [2],
         "count": [2],
-        "list": _sort_list_column_data([[1, None, 2]]),
     }
 
     daft_df.collect()
     daft_cols = daft_df.to_pydict()
-    daft_cols = {**daft_cols, "list": _sort_list_column_data(daft_cols["list"])}
 
-    assert_pydict_equals(
-        daft_cols,
-        expected,
-        assert_ordering=True,
-    )
+    assert pa.Table.from_pydict(daft_cols) == pa.Table.from_pydict(expected)
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 4])
@@ -69,7 +57,6 @@ def test_agg_global_all_null(repartition_nparts):
                 (col("values").alias("min"), "min"),
                 (col("values").alias("max"), "max"),
                 (col("values").alias("count"), "count"),
-                (col("values").alias("list"), "list"),
             ]
         )
     )
@@ -79,33 +66,12 @@ def test_agg_global_all_null(repartition_nparts):
         "min": [None],
         "max": [None],
         "count": [0],
-        "list": [[None, None, None]],
     }
 
     daft_df.collect()
     daft_cols = daft_df.to_pydict()
-    daft_cols = {**daft_cols, "list": _sort_list_column_data(daft_cols["list"])}
 
-    assert_pydict_equals(
-        daft_cols,
-        expected,
-        assert_ordering=True,
-    )
-
-
-def test_agg_global_null_type_column():
-    daft_df = DataFrame.from_pydict(
-        {
-            "id": [1, 2, 3],
-            "values": [None, None, None],
-        }
-    )
-    with pytest.raises(ExpressionTypeError):
-        daft_df.agg(
-            [
-                (col("values").alias("sum"), "sum"),
-            ]
-        )
+    assert pa.Table.from_pydict(daft_cols) == pa.Table.from_pydict(expected)
 
 
 def test_agg_global_empty():
@@ -125,27 +91,21 @@ def test_agg_global_empty():
                 (col("values").alias("min"), "min"),
                 (col("values").alias("max"), "max"),
                 (col("values").alias("count"), "count"),
-                (col("values").alias("list"), "list"),
             ]
         )
     )
     expected = {
-        "sum": [],
-        "mean": [],
-        "min": [],
-        "max": [],
-        "count": [],
-        "list": [],
+        "sum": [None],
+        "mean": [None],
+        "min": [None],
+        "max": [None],
+        "count": [0],
     }
 
     daft_df.collect()
     daft_cols = daft_df.to_pydict()
 
-    assert_pydict_equals(
-        daft_cols,
-        expected,
-        assert_ordering=True,
-    )
+    assert pa.Table.from_pydict(daft_cols) == pa.Table.from_pydict(expected)
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 7])
@@ -166,7 +126,6 @@ def test_agg_groupby(repartition_nparts):
                 (col("values").alias("min"), "min"),
                 (col("values").alias("max"), "max"),
                 (col("values").alias("count"), "count"),
-                (col("values").alias("list"), "list"),
             ]
         )
     )
@@ -177,18 +136,12 @@ def test_agg_groupby(repartition_nparts):
         "min": [1, 2],
         "max": [2, 4],
         "count": [2, 2],
-        "list": _sort_list_column_data([[1, None, 2], [2, None, 4]]),
     }
 
     daft_df.collect()
     daft_cols = daft_df.to_pydict()
-    daft_cols = {**daft_cols, "list": _sort_list_column_data(daft_cols["list"])}
 
-    assert_pydict_equals(
-        daft_cols,
-        expected,
-        sort_key="group",
-    )
+    assert pa.Table.from_pydict(daft_cols).sort_by("group") == pa.Table.from_pydict(expected).sort_by("group")
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 5])
@@ -209,7 +162,6 @@ def test_agg_groupby_all_null(repartition_nparts):
             (col("values").alias("min"), "min"),
             (col("values").alias("max"), "max"),
             (col("values").alias("count"), "count"),
-            (col("values").alias("list"), "list"),
         ]
     )
 
@@ -220,17 +172,12 @@ def test_agg_groupby_all_null(repartition_nparts):
         "min": [None, None],
         "max": [None, None],
         "count": [0, 0],
-        "list": [[None, None], [None, None]],
     }
 
     daft_df.collect()
     daft_cols = daft_df.to_pydict()
 
-    assert_pydict_equals(
-        daft_cols,
-        expected,
-        sort_key="group",
-    )
+    assert pa.Table.from_pydict(daft_cols).sort_by("group") == pa.Table.from_pydict(expected).sort_by("group")
 
 
 def test_agg_groupby_null_type_column():
@@ -243,7 +190,7 @@ def test_agg_groupby_null_type_column():
     )
     daft_df = daft_df.groupby(col("group"))
 
-    with pytest.raises(ExpressionTypeError):
+    with pytest.raises(ValueError):
         daft_df.agg(
             [
                 (col("values").alias("sum"), "sum"),
@@ -275,8 +222,8 @@ def test_null_groupby_keys(repartition_nparts):
         "group": [0, 1, 2, None],
         "mean": [0.0, 1.0, 2.0, 3.0],
     }
-    daft_df.collect()
-    assert_pydict_equals(daft_df.to_pydict(), expected, sort_key="group")
+    daft_cols = daft_df.to_pydict()
+    assert pa.Table.from_pydict(daft_cols).sort_by("group") == pa.Table.from_pydict(expected).sort_by("group")
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 4])
@@ -291,7 +238,7 @@ def test_all_null_groupby_keys(repartition_nparts):
 
     daft_df = (
         daft_df.repartition(repartition_nparts)
-        .with_column("group", daft_df["group"].cast(int))
+        .with_column("group", daft_df["group"].cast(DataType.int64()))
         .groupby(col("group"))
         .agg(
             [
@@ -304,8 +251,8 @@ def test_all_null_groupby_keys(repartition_nparts):
         "group": [None],
         "mean": [2.0],
     }
-    daft_df.collect()
-    assert_pydict_equals(daft_df.to_pydict(), expected, sort_key="group")
+    daft_cols = daft_df.to_pydict()
+    assert pa.Table.from_pydict(daft_cols).sort_by("group") == pa.Table.from_pydict(expected).sort_by("group")
 
 
 def test_null_type_column_groupby_keys():
@@ -338,7 +285,6 @@ def test_agg_groupby_empty():
             (col("values").alias("min"), "min"),
             (col("values").alias("max"), "max"),
             (col("values").alias("count"), "count"),
-            (col("values").alias("list"), "list"),
         ]
     )
 
@@ -349,14 +295,8 @@ def test_agg_groupby_empty():
         "min": [],
         "max": [],
         "count": [],
-        "list": [],
     }
 
     daft_df.collect()
     daft_cols = daft_df.to_pydict()
-
-    assert_pydict_equals(
-        daft_cols,
-        expected,
-        sort_key="group",
-    )
+    assert pa.Table.from_pydict(daft_cols).sort_by("group") == pa.Table.from_pydict(expected).sort_by("group")
