+++ tests/dataframe/test_joins.py
@@ -1,21 +1,62 @@
 from __future__ import annotations
 
+import pyarrow as pa
 import pytest
 
-from daft import DataFrame
+import daft
+from daft.datatype import DataType
 from daft.errors import ExpressionTypeError
-from tests.conftest import assert_pydict_equals
+
+
+@pytest.mark.parametrize("n_partitions", [1, 2, 4])
+def test_multicol_joins(n_partitions: int):
+    df = daft.DataFrame.from_pydict(
+        {
+            "A": [1, 2, 3],
+            "B": ["a", "b", "c"],
+            "C": [True, False, True],
+        }
+    ).repartition(n_partitions, "A", "B")
+
+    joined = df.join(df, on=["A", "B"]).sort("A")
+    joined_data = joined.to_pydict()
+
+    assert joined_data == {
+        "A": [1, 2, 3],
+        "B": ["a", "b", "c"],
+        "C": [True, False, True],
+        "right.C": [True, False, True],
+    }
+
+
+@pytest.mark.parametrize("n_partitions", [1, 2, 4])
+def test_limit_after_join(n_partitions: int):
+    data = {
+        "A": [1, 2, 3],
+    }
+    df1 = daft.DataFrame.from_pydict(data).repartition(n_partitions, "A")
+    df2 = daft.DataFrame.from_pydict(data).repartition(n_partitions, "A")
+
+    joined = df1.join(df2, on="A").limit(1)
+    joined_data = joined.to_pydict()
+    assert "A" in joined_data
+    assert len(joined_data["A"]) == 1
+
+
+###
+# Tests for nulls
+###
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 4])
 def test_inner_join(repartition_nparts):
-    daft_df = DataFrame.from_pydict(
+    daft_df = daft.DataFrame.from_pydict(
         {
             "id": [1, None, 3],
             "values_left": ["a1", "b1", "c1"],
         }
     ).repartition(repartition_nparts)
-    daft_df2 = DataFrame.from_pydict(
+    daft_df2 = daft.DataFrame.from_pydict(
         {
             "id": [1, 2, 3],
             "values_right": ["a2", "b2", "c2"],
@@ -28,21 +69,19 @@ def test_inner_join(repartition_nparts):
         "values_left": ["a1", "c1"],
         "values_right": ["a2", "c2"],
     }
-    daft_df.collect()
-
-    assert_pydict_equals(daft_df.to_pydict(), expected, sort_key="id")
+    assert pa.Table.from_pydict(daft_df.to_pydict()).sort_by("id") == pa.Table.from_pydict(expected).sort_by("id")
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 4])
 def test_inner_join_multikey(repartition_nparts):
-    daft_df = DataFrame.from_pydict(
+    daft_df = daft.DataFrame.from_pydict(
         {
             "id": [1, None, None],
             "id2": ["foo1", "foo2", None],
             "values_left": ["a1", "b1", "c1"],
         }
     ).repartition(repartition_nparts)
-    daft_df2 = DataFrame.from_pydict(
+    daft_df2 = daft.DataFrame.from_pydict(
         {
             "id": [None, None, 1],
             "id2": ["foo2", None, "foo1"],
@@ -57,43 +96,41 @@ def test_inner_join_multikey(repartition_nparts):
         "values_left": ["a1"],
         "values_right": ["c2"],
     }
-    daft_df.collect()
-    assert_pydict_equals(daft_df.to_pydict(), expected, sort_key="id")
+    assert pa.Table.from_pydict(daft_df.to_pydict()).sort_by("id") == pa.Table.from_pydict(expected).sort_by("id")
 
 
 @pytest.mark.parametrize("repartition_nparts", [1, 2, 4])
 def test_inner_join_all_null(repartition_nparts):
-    daft_df = DataFrame.from_pydict(
+    daft_df = daft.DataFrame.from_pydict(
         {
             "id": [None, None, None],
             "values_left": ["a1", "b1", "c1"],
         }
     ).repartition(repartition_nparts)
-    daft_df2 = DataFrame.from_pydict(
+    daft_df2 = daft.DataFrame.from_pydict(
         {
             "id": [1, 2, 3],
             "values_right": ["a2", "b2", "c2"],
         }
     ).repartition(repartition_nparts)
-    daft_df = daft_df.with_column("id", daft_df["id"].cast(int)).join(daft_df2, on="id", how="inner")
+    daft_df = daft_df.with_column("id", daft_df["id"].cast(DataType.int64())).join(daft_df2, on="id", how="inner")
 
     expected = {
         "id": [],
         "values_left": [],
         "values_right": [],
     }
-    daft_df.collect()
-    assert_pydict_equals(daft_df.to_pydict(), expected, sort_key="id")
+    assert pa.Table.from_pydict(daft_df.to_pydict()).sort_by("id") == pa.Table.from_pydict(expected).sort_by("id")
 
 
 def test_inner_join_null_type_column():
-    daft_df = DataFrame.from_pydict(
+    daft_df = daft.DataFrame.from_pydict(
         {
             "id": [None, None, None],
             "values_left": ["a1", "b1", "c1"],
         }
     )
-    daft_df2 = DataFrame.from_pydict(
+    daft_df2 = daft.DataFrame.from_pydict(
         {
             "id": [None, None, None],
             "values_right": ["a2", "b2", "c2"],
