--- tests_legacy/assets/311-service-requests.24.partial_empty/0-12.csv
@@ -1,13 +1,0 @@
-Unique Key,Created Date,Closed Date,Agency,Agency Name,Complaint Type,Descriptor,Location Type,Incident Zip,Incident Address,Street Name,Cross Street 1,Cross Street 2,Intersection Street 1,Intersection Street 2,Address Type,City,Landmark,Facility Type,Status,Due Date,Resolution Action Updated Date,Community Board,Borough,X Coordinate (State Plane),Y Coordinate (State Plane),Park Facility Name,Park Borough,School Name,School Number,School Region,School Code,School Phone Number,School Address,School City,School State,School Zip,School Not Found,School or Citywide Complaint,Vehicle Type,Taxi Company Borough,Taxi Pick Up Location,Bridge Highway Name,Bridge Highway Direction,Road Ramp,Bridge Highway Segment,Garage Lot Name,Ferry Direction,Ferry Terminal Name,Latitude,Longitude,Location
-26589651,10/31/2013 02:08:41 AM,,NYPD,New York City Police Department,Noise - Street/Sidewalk,Loud Talking,Street/Sidewalk,11432,90-03 169 STREET,169 STREET,90 AVENUE,91 AVENUE,,,ADDRESS,JAMAICA,,Precinct,Assigned,10/31/2013 10:08:41 AM,10/31/2013 02:35:17 AM,12 QUEENS,QUEENS,1042027,197389,Unspecified,QUEENS,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.70827532593202,-73.79160395779721,"(40.70827532593202, -73.79160395779721)"
-26593698,10/31/2013 02:01:04 AM,,NYPD,New York City Police Department,Illegal Parking,Commercial Overnight Parking,Street/Sidewalk,11378,58 AVENUE,58 AVENUE,58 PLACE,59 STREET,,,BLOCKFACE,MASPETH,,Precinct,Open,10/31/2013 10:01:04 AM,,05 QUEENS,QUEENS,1009349,201984,Unspecified,QUEENS,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.721040535628305,-73.90945306791765,"(40.721040535628305, -73.90945306791765)"
-26594139,10/31/2013 02:00:24 AM,10/31/2013 02:40:32 AM,NYPD,New York City Police Department,Noise - Commercial,Loud Music/Party,Club/Bar/Restaurant,10032,4060 BROADWAY,BROADWAY,WEST 171 STREET,WEST 172 STREET,,,ADDRESS,NEW YORK,,Precinct,Closed,10/31/2013 10:00:24 AM,10/31/2013 02:39:42 AM,12 MANHATTAN,MANHATTAN,1001088,246531,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.84332975466513,-73.93914371913482,"(40.84332975466513, -73.93914371913482)"
-26595721,10/31/2013 01:56:23 AM,10/31/2013 02:21:48 AM,NYPD,New York City Police Department,Noise - Vehicle,Car/Truck Horn,Street/Sidewalk,10023,WEST 72 STREET,WEST 72 STREET,COLUMBUS AVENUE,AMSTERDAM AVENUE,,,BLOCKFACE,NEW YORK,,Precinct,Closed,10/31/2013 09:56:23 AM,10/31/2013 02:21:10 AM,07 MANHATTAN,MANHATTAN,989730,222727,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.7780087446372,-73.98021349023975,"(40.7780087446372, -73.98021349023975)"
-26590930,10/31/2013 01:53:44 AM,,DOHMH,Department of Health and Mental Hygiene,Rodent,Condition Attracting Rodents,Vacant Lot,10027,WEST 124 STREET,WEST 124 STREET,LENOX AVENUE,ADAM CLAYTON POWELL JR BOULEVARD,,,BLOCKFACE,NEW YORK,,N/A,Pending,11/30/2013 01:53:44 AM,10/31/2013 01:59:54 AM,10 MANHATTAN,MANHATTAN,998815,233545,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.80769092704951,-73.94738703491433,"(40.80769092704951, -73.94738703491433)"
-26592370,10/31/2013 01:46:52 AM,,NYPD,New York City Police Department,Noise - Commercial,Banging/Pounding,Club/Bar/Restaurant,11372,37 AVENUE,37 AVENUE,84 STREET,85 STREET,,,BLOCKFACE,JACKSON HEIGHTS,,Precinct,Open,10/31/2013 09:46:52 AM,,03 QUEENS,QUEENS,1016948,212540,Unspecified,QUEENS,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.7499893014072,-73.88198770727831,"(40.7499893014072, -73.88198770727831)"
-26595682,10/31/2013 01:46:40 AM,,NYPD,New York City Police Department,Blocked Driveway,No Access,Street/Sidewalk,11419,107-50 109 STREET,109 STREET,107 AVENUE,109 AVENUE,,,ADDRESS,SOUTH RICHMOND HILL,,Precinct,Assigned,10/31/2013 09:46:40 AM,10/31/2013 01:59:51 AM,10 QUEENS,QUEENS,1030919,187622,Unspecified,QUEENS,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.68153278675525,-73.83173699701601,"(40.68153278675525, -73.83173699701601)"
-26595195,10/31/2013 01:44:19 AM,10/31/2013 01:58:49 AM,NYPD,New York City Police Department,Noise - Commercial,Loud Music/Party,Club/Bar/Restaurant,11417,137-09 CROSSBAY BOULEVARD,CROSSBAY BOULEVARD,PITKIN AVENUE,VAN WICKLEN ROAD,,,ADDRESS,OZONE PARK,,Precinct,Closed,10/31/2013 09:44:19 AM,10/31/2013 01:58:49 AM,10 QUEENS,QUEENS,1027776,184076,Unspecified,QUEENS,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.67181584567338,-73.84309181950769,"(40.67181584567338, -73.84309181950769)"
-26590540,10/31/2013 01:44:14 AM,10/31/2013 02:28:04 AM,NYPD,New York City Police Department,Noise - Commercial,Loud Talking,Club/Bar/Restaurant,10011,258 WEST 15 STREET,WEST 15 STREET,7 AVENUE,8 AVENUE,,,ADDRESS,NEW YORK,,Precinct,Closed,10/31/2013 09:44:14 AM,10/31/2013 02:00:56 AM,04 MANHATTAN,MANHATTAN,984031,208847,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.73991339303542,-74.00079028612932,"(40.73991339303542, -74.00079028612932)"
-26594392,10/31/2013 01:34:41 AM,10/31/2013 02:23:51 AM,NYPD,New York City Police Department,Noise - Commercial,Loud Music/Party,Club/Bar/Restaurant,11225,835 NOSTRAND AVENUE,NOSTRAND AVENUE,UNION STREET,PRESIDENT STREET,,,ADDRESS,BROOKLYN,,Precinct,Closed,10/31/2013 09:34:41 AM,10/31/2013 01:48:26 AM,09 BROOKLYN,BROOKLYN,997941,182725,Unspecified,BROOKLYN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.66820406598287,-73.95064760056546,"(40.66820406598287, -73.95064760056546)"
-26595176,10/31/2013 01:25:12 AM,,NYPD,New York City Police Department,Noise - House of Worship,Loud Music/Party,House of Worship,11218,3775 18 AVENUE,18 AVENUE,EAST 9 STREET,EAST 8 STREET,,,ADDRESS,BROOKLYN,,Precinct,Open,10/31/2013 09:25:12 AM,,14 BROOKLYN,BROOKLYN,992726,170399,Unspecified,BROOKLYN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.63437840816299,-73.96946177104543,"(40.63437840816299, -73.96946177104543)"
-26591982,10/31/2013 01:24:14 AM,10/31/2013 01:54:39 AM,NYPD,New York City Police Department,Noise - Commercial,Loud Music/Party,Club/Bar/Restaurant,10003,187 2 AVENUE,2 AVENUE,EAST 11 STREET,EAST 12 STREET,,,ADDRESS,NEW YORK,,Precinct,Closed,10/31/2013 09:24:14 AM,10/31/2013 01:54:39 AM,03 MANHATTAN,MANHATTAN,988110,205533,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.73081644089586,-73.98607265739876,"(40.73081644089586, -73.98607265739876)"
