--- tests_legacy/assets/311-service-requests.24.partial_empty/12-24.csv
@@ -1,13 +1,0 @@
-Unique Key,Created Date,Closed Date,Agency,Agency Name,Complaint Type,Descriptor,Location Type,Incident Zip,Incident Address,Street Name,Cross Street 1,Cross Street 2,Intersection Street 1,Intersection Street 2,Address Type,City,Landmark,Facility Type,Status,Due Date,Resolution Action Updated Date,Community Board,Borough,X Coordinate (State Plane),Y Coordinate (State Plane),Park Facility Name,Park Borough,School Name,School Number,School Region,School Code,School Phone Number,School Address,School City,School State,School Zip,School Not Found,School or Citywide Complaint,Vehicle Type,Taxi Company Borough,Taxi Pick Up Location,Bridge Highway Name,Bridge Highway Direction,Road Ramp,Bridge Highway Segment,Garage Lot Name,Ferry Direction,Ferry Terminal Name,Latitude,Longitude,Location
-26591573,10/31/2013 12:35:18 AM,10/31/2013 02:41:35 AM,NYPD,New York City Police Department,Noise - Street/Sidewalk,Loud Talking,Street/Sidewalk,10312,24 PRINCETON LANE,PRINCETON LANE,HAMPTON GREEN,DEAD END,,,ADDRESS,STATEN ISLAND,,Precinct,Closed,10/31/2013 08:35:18 AM,10/31/2013 01:45:17 AM,03 STATEN ISLAND,STATEN ISLAND,929577,140964,Unspecified,STATEN ISLAND,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.55342078716953,-74.19674315017886,"(40.55342078716953, -74.19674315017886)"
-26590509,10/31/2013 12:33:00 AM,,DOT,Department of Transportation,Street Light Condition,Street Light Out,,,38 ST E,38 ST E,CHURCH AVE,LINDEN BLVD,,,BLOCKFACE,,,N/A,Open,,,Unspecified BROOKLYN,BROOKLYN,,,Unspecified,BROOKLYN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,,,,,,,,,,,,,,,
-26591379,10/31/2013 12:32:44 AM,,DOHMH,Department of Health and Mental Hygiene,Harboring Bees/Wasps,Bees/Wasps - Not a beekeper,3+ Family Mixed Use Building,10025,501 WEST 110 STREET,WEST 110 STREET,AMSTERDAM AVENUE,BROADWAY,,,ADDRESS,NEW YORK,,N/A,Open,11/30/2013 12:32:44 AM,,09 MANHATTAN,MANHATTAN,994143,231888,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.80314938553783,-73.96426608076082,"(40.80314938553783, -73.96426608076082)"
-26594085,10/31/2013 12:32:08 AM,,NYPD,New York City Police Department,Noise - Street/Sidewalk,Loud Talking,Street/Sidewalk,10026,121 WEST 116 STREET,WEST 116 STREET,LENOX AVENUE,7 AVENUE,,,ADDRESS,NEW YORK,,Precinct,Assigned,10/31/2013 08:32:08 AM,10/31/2013 02:00:57 AM,10 MANHATTAN,MANHATTAN,997947,231613,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.80238950799943,-73.95052644123253,"(40.80238950799943, -73.95052644123253)"
-26589201,10/31/2013 12:32:00 AM,,DOT,Department of Transportation,Street Light Condition,Street Light Out,,10309,295 BAYVIEW AVENUE,BAYVIEW AVENUE,VAIL AVENUE,BAYVIEW LANE,,,ADDRESS,STATEN ISLAND,,N/A,Open,,,03 STATEN ISLAND,STATEN ISLAND,927687,127837,Unspecified,STATEN ISLAND,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,,,,,,,,,,,,,40.517377871705676,-74.20343466779575,"(40.517377871705676, -74.20343466779575)"
-26591641,10/31/2013 12:31:17 AM,10/31/2013 02:41:36 AM,NYPD,New York City Police Department,Blocked Driveway,No Access,Street/Sidewalk,10312,24 PRINCETON LANE,PRINCETON LANE,HAMPTON GREEN,DEAD END,,,ADDRESS,STATEN ISLAND,,Precinct,Closed,10/31/2013 08:31:17 AM,10/31/2013 01:43:09 AM,03 STATEN ISLAND,STATEN ISLAND,929577,140964,Unspecified,STATEN ISLAND,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.55342078716953,-74.19674315017886,"(40.55342078716953, -74.19674315017886)"
-26595564,10/31/2013 12:30:36 AM,,NYPD,New York City Police Department,Noise - Street/Sidewalk,Loud Music/Party,Street/Sidewalk,11236,AVENUE J,AVENUE J,EAST 80 STREET,EAST 81 STREET,,,BLOCKFACE,BROOKLYN,,Precinct,Open,10/31/2013 08:30:36 AM,,18 BROOKLYN,BROOKLYN,1008937,170310,Unspecified,BROOKLYN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.634103775951736,-73.91105541883589,"(40.634103775951736, -73.91105541883589)"
-26591378,10/31/2013 12:30:31 AM,,TLC,Taxi and Limousine Commission,Taxi Complaint,Driver Complaint,,10036,645 10 AVENUE,10 AVENUE,WEST 45 STREET,WEST 46 STREET,,,ADDRESS,NEW YORK,,N/A,Open,,,04 MANHATTAN,MANHATTAN,985965,216868,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,Other,,,,,,,,40.761928847500016,-73.99380918401052,"(40.761928847500016, -73.99380918401052)"
-26593872,10/31/2013 12:29:47 AM,10/31/2013 12:38:29 AM,NYPD,New York City Police Department,Noise - House of Worship,Banging/Pounding,House of Worship,10025,,,,,WEST   99 STREET,AMSTERDAM AVENUE,INTERSECTION,NEW YORK,,Precinct,Closed,10/31/2013 08:29:47 AM,10/31/2013 12:38:29 AM,07 MANHATTAN,MANHATTAN,992846,229279,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.795989749917204,-73.96895423714467,"(40.795989749917204, -73.96895423714467)"
-26591420,10/31/2013 12:28:30 AM,10/31/2013 02:06:11 AM,NYPD,New York City Police Department,Homeless Encampment,N/A,Residential Building/House,10025,2754 BROADWAY,BROADWAY,WEST 105 STREET,WEST 106 STREET,,,ADDRESS,NEW YORK,,Precinct,Closed,10/31/2013 08:28:30 AM,10/31/2013 02:06:11 AM,07 MANHATTAN,MANHATTAN,993139,231139,Unspecified,MANHATTAN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.8010946529914,-73.96789356094007,"(40.8010946529914, -73.96789356094007)"
-26592976,10/31/2013 12:23:24 AM,10/31/2013 01:05:41 AM,NYPD,New York City Police Department,Blocked Driveway,No Access,Street/Sidewalk,11433,173-41 103 ROAD,103 ROAD,173 STREET,177 STREET,,,ADDRESS,JAMAICA,,Precinct,Closed,10/31/2013 08:23:24 AM,10/31/2013 01:05:41 AM,12 QUEENS,QUEENS,1044124,195866,Unspecified,QUEENS,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,N,,,,,,,,,,,,40.70408112125158,-73.78405385422116,"(40.70408112125158, -73.78405385422116)"
-26590262,10/31/2013 12:23:00 AM,,DOT,Department of Transportation,Traffic Signal Condition,Controller,,11235,,,,,SHORE BOULEVARD,CASS PLACE,INTERSECTION,BROOKLYN,,N/A,Open,,,15 BROOKLYN,BROOKLYN,997073,151225,Unspecified,BROOKLYN,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,Unspecified,,,,,,,,,,,,,40.5817444882428,-73.95383634845487,"(40.5817444882428, -73.95383634845487)"
